Included fully cross platform photo selector and barcode reader(untest reader), have fixed some small display issues and now have the device fully comitting to database for items
Related Work Items: #5

+++ Plutus/Plutus/Plutus/Models/ItemModel.cs
@@ -13,8 +13,8 @@ namespace Plutus.Models
         public string Name { get; set; }
         public string Brand { get; set; }
         public string Desc { get; set; }
-        public double Cost { get; set; }
-        public double Price { get; set; }
+        public decimal Cost { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
         public byte[] Image { get; set; }
